Add files via upload
RFID Based Security System

Radio Frequency Identification (RFID) is a technology that uses electromagnetic fields to automatically identify and track tags attached to objects. The tags contain electronically stored information that can be read from a distance without direct line-of-sight. RFID systems typically consist of a reader and tags; the reader emits radio waves and receives signals back from the tags, which can either be passive (powered by the reader's signal) or active (battery-powered).

Project Explanation:

My project involves developing an advanced security system utilizing RFID technology and an 8051 microcontroller. The system features an RFID reader module that reads the card number from an RFID card and transmits it to the 8051 microcontroller via UART communication. Upon receiving the card number, the microcontroller verifies its validity against a stored list of authorized IDs. If the card number is valid, the system prompts the user to enter a password using a keypad. If the correct password is entered, the microcontroller activates a motor to rotate and open the gate; if the password is incorrect, a buzzer sounds to indicate access denial. This project showcases an effective security solution by integrating RFID technology and password authentication to manage secure access in environments such as offices and restricted areas.